Refactored the packages of the application to better match the advice given on reddit by /u/theothertomelliott
